+++ projects/world/tree/SimpleTreeDecorator.cpp
@@ -10,7 +10,8 @@
 namespace world {
 
 SimpleTreeDecorator::SimpleTreeDecorator(int maxTreesPerChunk)
-        : _maxTreesPerChunk(maxTreesPerChunk), _rng(time(NULL)) {
+        : _maxTreesPerChunk(maxTreesPerChunk),
+          _rng(static_cast<u32>(time(NULL))) {
 
     auto &skeletton = _model.addWorker<TreeSkelettonGenerator>();
     skeletton.setRootWeight(TreeParamsd::gaussian(3, 0.2));
